glocalfilemonitor: Make start () fallible
There are cases where starting a file monitor fails; in particular this
happens if the specific file system doesn't support the kernel-specific
file change notification mechanism (such as inotify or kqueue). This can
only be detected and reported at start () time, not at is_supported ()
time, since the latter is not passed the specific file to watch, and
this can differ from filesystem to filesystem.

Make it possible for start () to fail with an error, in which case
g_local_file_monitor_new_for_path () and
g_local_file_monitor_new_in_worker () will return NULL and propagate
the error, as they already do when no system-specific GLocalFileMonitor
subclass has been found. This will cause g_file_monitor_file () to fall
back to a GPollFileMonitor.

The existing GLocalFileMonitor subclasses are adapted to the API change,
and GWin32FileMonitor makes an attempt to actually use it to propagate
errors.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>