gthemedicon: add style variant fallback only if use-default-fallbacks
Add the style variant (symbolic/regular) only if use-default-fallbacks
is true, since the style variant is also a kind of fallback.

When use-default-fallbacks is false it's still possible to control
manually the list of icons to lookup, of course. Or even to create
two GThemedIcon instances to perform a "full" lookup before trying the
style variant.

Fixes https://gitlab.gnome.org/GNOME/glib/-/issues/3376