gmarkup: make documentation more discoverable
There is a bunch of documentation in a separate markdown page that does
not appear in search results. We should point to it.

I think the restriction on not being used to process untrusted input
should be relaxed into something more permissive, like "should not be
used on untrusted input if you care about denial of service." But that
can be a problem for another day.