gitypeinfo: Add serialization and deserialization API
Adds three new functions:
- gi_type_info_serialize()
- gi_repository_new_type_info_from_bytes()
- gi_repository_load_type_info_from_bytes()

This allows writing the offset and flags of a GITypeInfo in a byte buffer,
which can be stored in a memory-efficient way. The GITypeInfo can then be
created on demand from the byte buffer, the repository, the typelib, and
the container info. The repository, typelib, or container may be able to
be elided by the API consumer, depending on the application if they are
already obtainable somewhere else, so a GITypeInfo could potentially be
"dehydrated" into only 5 bytes.

The bytes are only intended to be used within the same process.

The buffer length is defined in a constant macro, so that callers can
declare an array on the stack. But it does not need to stay constant
across versions of GIRepository.