tests: Update desktop-files with an intent-aware update-desktop-database
The result is that we get the `intent.cache` cache for fast lookups of
intents. In the next commit, we will switch over the implementation for
finding apps which implement interfaces to the new one which uses
`intentapps.list` and `intent.cache` instead of iterating over all the
desktop files.