gdesktopappinfo: Add support for the intentapps specification
Add support for the intentapps-spec, inspired by the mimeapps-spec.

The specification describes directories to look for `intentapps.list`
files, the format for this file and relevant fields in the desktop
entry.

The new functions takes an interface and an optional scope. The
interfaces correspond to the `Implements` key and the scope may be used
by interfaces to define different default applications for a specific
context.