gdesktopappinfo: Use intent org.freedesktop.Terminal1 for terminal apps
We now have everything in place to quickly look up the default app which
implements the org.freedesktop.Terminal1 intent. We can then call the
dbus interface on that app to launch desktop files which require a
terminal ro run, instead of exec-ing a terminal from the hard-coded list
we currently have.