gatomic/MSVC: Use more Interlocked wrappers from winnt.h
The Interlocked functions from winnt.h are mostly wrappers around
the compiler's _Interlocked intrinsics (MSVC, Windows SDK) or
__sync builtins (GCC, mingw-w64).

Now, historically the Windows SDK did not provide a few Interlocked
functions on x86, for example InterlockedAnd, InterlockedOr, and
InterlockedXor. In fact the MSDN documentation suggests using the
compiler intrinsics in this case [1]:

 "For the Intel Itanium-based systems and x64 architectures, this
  function is implemented using the compiler intrinsic. For the x86
  architecture, use the compiler _InterlockedAnd intrinsic directly"

However these functions are implemented also for x86 starting with
the Windows 8.1 SDK. Let's fallback to the intrinsics for VS < 2015.

References:

 1. https://learn.microsoft.com/en-us/windows/win32/api/winnt/nf-winnt-interlockedand
 2. https://bugzilla.gnome.org/show_bug.cgi?id=652000